requirements.txt
playwright
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
webvtt-py
google-generativeai